Greeting message displayed only in echo area
The message-box displays the message in a box, that has a button. It
is required to be pressed so the load of the configuration would go
on. That makes automatic start of Emacs demand human interaction.

Not needing interaction makes it possible to start the editor on boot
more fluently.

Signed-off-by: m00ns7ruck <m00ns7ruck@abv.bg>